refactoring header component
