+++ src/components/Header.jsx
@@ -1,8 +1,7 @@
 import { GiHamburgerMenu } from "react-icons/gi";
-import { IoClose } from "react-icons/io5";
 import { Navigation } from "./Navigation";
-import { SocialLinks } from "./SocialLinks";
 import logo from "../assets/image/logo.webp";
+import { MobileMenu } from "./MobileMenu";
 
 const Header = () => {
   // function to open menu for mobile
@@ -10,14 +9,6 @@ const Header = () => {
     const openMenu = document.querySelector(".mobile-menu");
     if (openMenu.style.display === "none") {
       openMenu.style.display = "block";
-    }
-  };
-
-  // function to close menu for mobile
-  const handleCloseMenu = () => {
-    const openMenu = document.querySelector(".mobile-menu");
-    if (openMenu.style.display === "block") {
-      openMenu.style.display = "none";
     }
   };
 
@@ -32,20 +23,7 @@ const Header = () => {
           <Navigation marginLeft="30px" display="flex" />
         </div>
       </div>
-      <div className="mobile-menu" style={{ display: "none" }}>
-        <div className="close-menu" onClick={handleCloseMenu}>
-          <IoClose className="close-menu-icon" />
-        </div>
-        <div className="mobile-navigation">
-          <Navigation
-            flexDirection="column"
-            alingItems="flex-start"
-            marginBottom="20px"
-            color="#FFFFFF"
-          />
-        </div>
-        <SocialLinks justifyContent="flex-start" />
-      </div>
+      <MobileMenu />
     </header>
   );
 };
